requirements.txt
coverage==4.5.3
Django==2.2.2
djangorestframework==3.9.4
pkg-resources==0.0.0
pytz==2019.1
sqlparse==0.3.0
Cython==0.25.2
Kivy==1.10.1
